Make coverage.sh work with dune
When I've converted nbd to dune, coverage.sh started failing with this
error:
From the command line:
Error: Don't know about directory .coverage!

Apparently dune doesn't like it when we nest a project inside itself. So
I've changed the coverage directory to be a temporary dir in /tmp - this
works fine with dune.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>